+++ packages/primitive/deco-object/index.js
@@ -29,17 +29,17 @@ const { object } = Node.prototype
  */
 export function DecoObject(conf) {
   conf = conf ?? this ?? {}
-  conf.pres = parsePresm(conf?.pres ?? conf, PRES)
+  const presm = parsePresm(conf?.pres ?? conf, PRES)
   const thres = conf.thres ?? 0, indent = conf.indent, surge = conf.surge
-  const proc = object.bind(new Node(conf))
+  const proc = object.bind(Node.init(presm))
   return (o, thr, ind, sur) => proc(o, thr ?? thres, ind ?? indent, sur ?? surge)
 }
 
 export function decoObject(ent, thr, ind, sur) {
   const conf = this ?? {}
-  conf.pres = parsePresm(conf?.pres ?? conf, PRES)
+  const presm = parsePresm(conf?.pres ?? conf, PRES)
   const thres = conf.thres ?? 0, indent = conf.indent, surge = conf.surge
-  return object.call(new Node(conf), ent, thr ?? thres, ind ?? indent, sur ?? surge)
+  return object.call(Node.init(presm), ent, thr ?? thres, ind ?? indent, sur ?? surge)
 }
 
 export {