+++ packages/interop/deco-samples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @spare/deco-samples
 
+## 0.16.17
+
+### Patch Changes
+
+- add @spare/serial
+- Updated dependencies
+  - @spare/node@0.16.17
+  - @spare/serial@0.16.17
+
 ## 0.16.16
 
 ### Patch Changes