+++ packages/primitive/deco-string/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @spare/deco-string
 
+## 0.16.17
+
+### Patch Changes
+
+- add @spare/serial
+- Updated dependencies
+  - @spare/node@0.16.17
+
 ## 0.16.16
 
 ### Patch Changes